move exporter management in dedicated folder and give possibility to use custom exporter
